Fix propagating of configdn, configpw, monitordn, monitorpw into the slapd.conf file.
Also, be more cautious, when, especially the *pw variables get
written to the file. Further, the suffix= line in the database config
section is not allowed:

Basically fixing error messages like this with enabled config and monitoring
databases:

543e1cc9 /etc/openldap/slapd.conf: line 23: suffix <cn=config> not allowed in config database.
543e1cc9 /etc/openldap/slapd.conf: line 23: <suffix> handler exited with 1!

with rootpw under database config
543e1d10 line 25 (rootpw ***)
543e1d10 /etc/openldap/slapd.conf: line 25: <rootpw> can only be set when rootdn is under suffix

with rootpw under database monitor
543e1dc6 line 38 (rootpw ***)
543e1dc6 /etc/openldap/slapd.conf: line 38: <rootpw> can only be set when rootdn is under suffix

For that last reason, the suffix= line in the config database
is commented out, as it is done for the monitor database already.